RacingGame의 결과를 판정하는 심판(RacingGameJudge) 생성
